pyproject.toml
[tool.poetry]
name = "clip2tex"
version = "0.1.0"
description = ""
authors = ["zombie110year <zombie110year@outlook.com>"]
packages = [
    { include = "winclip" }
]

[tool.poetry.dependencies]
python = "^3.7"
pywin32 = "^225.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
